Se arregla PHPUnit para login y registro de usuarios

+++ database/migrations/2016_07_07_165949_cambiarCustodio.php
@@ -12,11 +12,13 @@ class CambiarCustodio extends Migration
      */
     public function up()
     {
-        Schema::table('custodios', function (Blueprint $table) {
-            $table->string('nombre_responsable');
-            $table->string('area_piso');
-            $table->dropColumn('area-piso');
-        });
+        if (env('DB_MIGRACIONES', 'false') == 'false') {
+            Schema::table('custodios', function (Blueprint $table) {
+                $table->string('nombre_responsable');
+                $table->string('area_piso');
+                $table->dropColumn('area-piso');
+            });
+        }
     }
 
     /**
